more modes finished
made the roam code work properly. there was a convenient code on unity API that actually found random points in a range so it was easy to implement.
Also, as an aside. If you're reading this I apologize for doing this so last minute. I've had to crunch for other classes and often do my class projects one at a time. I have gotten better sleep lately and apologize for coming into your classes and bordering on the edge unconsciousness. Your classes were great and the irony of falling asleep while you told us to get better sleep wasn't lost on me. I hope the rest of your day goes well.